move NavBar out of Room; minor clean up
